+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### CHANGES
 
 - `cmd/hls`: move threads param from args -> hls.yml config (pow_parallel, default=0->1)
-- `build/settings.yml`: move consumers_cap -> hls.yml config (qb_consumers, default=0->1)
+- `build/settings.yml`: move consumers_cap -> hls.yml config (qbp_consumers, default=0->1)
 
 <!-- ... -->